+++ dakara_server/playlist/models.py
@@ -127,10 +127,10 @@ class PlaylistEntry(OrderedModel):
         return player
 
 
-class KaraStatus(models.Model):
-    """Current status of the kara
+class Karaoke(models.Model):
+    """Current kara
 
-    The status is unique for now.
+    Unique for now.
     """
     STOP = "stop"
     PLAY = "play"
@@ -147,16 +147,20 @@ class KaraStatus(models.Model):
         default=PLAY,
         null=False,
     )
+    date_stop = models.DateTimeField(null=True)
 
     def __str__(self):
-        return "In {} mode".format(self.status)
+        return "In {} mode{}".format(
+            self.status,
+            " will stop at {}".format(self.date.stop) if self.date_stop else ""
+        )
 
     @classmethod
     def get_object(cls):
         """Get the first instance of kara status
         """
-        kara_status, _ = cls.objects.get_or_create(pk=1)
-        return kara_status
+        karaoke, _ = cls.objects.get_or_create(pk=1)
+        return karaoke
 
 
 class PlayerError(models.Model):