+++ dakara_server/playlist/permissions.py
@@ -4,7 +4,7 @@ from django.contrib.auth import get_user_model
 
 from users.permissions import BasePermissionCustom
 from library.models import Song
-from .models import PlaylistEntry, Player, KaraStatus
+from .models import PlaylistEntry, Player, Karaoke
 
 
 UserModel = get_user_model()
@@ -156,7 +156,7 @@ class IsPlayerOrReadOnly(BasePermissionCustom):
         return request.user.has_playlist_permission_level(UserModel.PLAYER)
 
 
-class KaraStatusIsNotStoppedOrReadOnly(permissions.BasePermission):
+class KaraokeIsNotStoppedOrReadOnly(permissions.BasePermission):
     """Grant access to not safe views if the kara is not in stop mode
     """
 
@@ -164,5 +164,5 @@ class KaraStatusIsNotStoppedOrReadOnly(permissions.BasePermission):
         if request.method in permissions.SAFE_METHODS:
             return True
 
-        kara_status = KaraStatus.get_object()
-        return kara_status.status != KaraStatus.STOP
+        karaoke = Karaoke.get_object()
+        return karaoke.status != Karaoke.STOP