+++ dakara_server/playlist/tests_playlist_entry.py
@@ -7,7 +7,7 @@ from django.contrib.auth import get_user_model
 from rest_framework import status
 
 from playlist.base_test import BaseAPITestCase, tz
-from playlist.models import PlaylistEntry, Player
+from playlist.models import PlaylistEntry, Player, Karaoke
 
 
 UserModel = get_user_model()
@@ -125,11 +125,11 @@ class PlaylistEntryListViewListCreateAPIViewTestCase(BaseAPITestCase):
         # Entry's owner is the user who created it
         self.assertEqual(new_entry.owner.id, self.p_user.id)
 
-    def test_post_create_playlist_entry_kara_status_stop_forbidden(self):
+    def test_post_create_playlist_entry_karaoke_stop_forbidden(self):
         """Test to verify playlist entry cannot be created when kara is stopped
         """
         # stop kara
-        self.set_kara_status_stop()
+        self.set_karaoke_stop()
 
         # Login as playlist user
         self.authenticate(self.manager)
@@ -156,6 +156,139 @@ class PlaylistEntryListViewListCreateAPIViewTestCase(BaseAPITestCase):
         # Post new playlist entry
         response = self.client.post(self.url, {"song_id": self.song1.id})
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
+        # post assert there are still 4 in database
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+    def test_post_create_playlist_entry_date_stop_success(self):
+        """Test user can add a song to playlist when before date stop
+        """
+        # set kara stop
+        date_stop = datetime.now(tz) + timedelta(hours=2)
+        karaoke = Karaoke.get_object()
+        karaoke.date_stop = date_stop
+        karaoke.save()
+
+        # login as user
+        self.authenticate(self.p_user)
+
+        # pre assert
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+        # request to add a new entry
+        response = self.client.post(self.url, {'song_id': self.song1.id})
+
+        # assert that the request is accepted
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(PlaylistEntry.objects.count(), 5)
+
+    def test_post_create_playlist_entry_date_stop_forbidden(self):
+        """Test user cannot add song to playlist after its date stop
+        """
+        # set kara stop
+        date_stop = datetime.now(tz)
+        karaoke = Karaoke.get_object()
+        karaoke.date_stop = date_stop
+        karaoke.save()
+
+        # login as user
+        self.authenticate(self.p_user)
+
+        # pre assert
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+        # request to add a new entry
+        response = self.client.post(self.url, {'song_id': self.song1.id})
+
+        # assert that the request is denied
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+    def test_post_create_playlist_entry_date_stop_success_manager(self):
+        """Test manager can add song to playlist after its date stop
+        """
+        # set kara stop
+        date_stop = datetime.now(tz)
+        karaoke = Karaoke.get_object()
+        karaoke.date_stop = date_stop
+        karaoke.save()
+
+        # login as manager
+        self.authenticate(self.manager)
+
+        # pre assert
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+        # request to add a new entry
+        response = self.client.post(self.url, {'song_id': self.song1.id})
+
+        # assert that the response and the database
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(PlaylistEntry.objects.count(), 5)
+
+    def test_post_create_playlist_entry_date_stop_success_admin(self):
+        """Test admin can add song to playlist after its date stop
+        """
+        # set kara stop
+        date_stop = datetime.now(tz)
+        karaoke = Karaoke.get_object()
+        karaoke.date_stop = date_stop
+        karaoke.save()
+
+        # login as admin
+        self.authenticate(self.admin)
+
+        # pre assert
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+        # request to add a new entry
+        response = self.client.post(self.url, {'song_id': self.song1.id})
+
+        # assert that the response and the database
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(PlaylistEntry.objects.count(), 5)
+
+    @patch('playlist.views.datetime',
+           side_effect=lambda *args, **kwargs: datetime(*args, **kwargs))
+    def test_post_create_playlist_entry_date_stop_forbidden_playlist_playing(
+            self, mocked_datetime):
+        """Test user cannot add song to playlist after its date stop
+
+        Test that only short enough songs can be added.
+        Test when the player is playing.
+        """
+        # patch the now method
+        now = datetime.now(tz)
+        mocked_datetime.now.return_value = now
+
+        # set the player
+        self.player_play_next_song(timing=timedelta(seconds=2))
+
+        # set kara stop such as to allow song1 to be added and not song2
+        date_stop = now + timedelta(seconds=20)
+        karaoke = Karaoke.get_object()
+        karaoke.date_stop = date_stop
+        karaoke.save()
+
+        # login as user
+        self.authenticate(self.p_user)
+
+        # pre assert
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+        # request to add a new entry which is too long
+        response = self.client.post(self.url, {'song_id': self.song2.id})
+
+        # assert that the request is denied
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+        # request to add a new entry which is short enough
+        response = self.client.post(self.url, {'song_id': self.song1.id})
+
+        # assert that the request is denied
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(PlaylistEntry.objects.count(), 5)
 
     def test_post_create_user_forbidden(self):
         """Test to verify simple user cannot create playlist entries