+++ CHANGELOG.md
@@ -34,6 +34,12 @@
 
 - Alternative titles of a work : search songs and works upon the alternative title of a work.
 - Reorder of the playlist
+- Prevent to add a song to the playlist after a certain karaoke stop date.
+
+### Changed
+
+- Kara status is now handled within the Karaoke object.
+- Karaoke route: `/playlist/kara-status/` > `/playlist/karaoke/`.
 
 ## 1.2.0 - 2018-06-03